Refactor project manipulation methods and add GraphMarkdownUtil for dependency graph generation
